requirements.txt
alembic==1.7.5
anyio==3.4.0
arrow==1.2.1
asgiref==3.4.1
Authlib==0.15.5
bcrypt==3.2.0
binaryornot==0.4.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==7.1.2
cookiecutter==1.7.3
cryptography==36.0.1
ecdsa==0.17.0
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
httptools==0.5.0
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
Jinja2==3.0.3
jinja2-time==0.2.0
Mako==1.1.6
MarkupSafe==2.0.1
mypy==1.2.0
mypy-extensions==1.0.0
passlib==1.7.4
poyo==0.5.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.5
python-slugify==5.0.2
PyYAML==6.0
requests==2.26.0
rsa==4.8
sentry-sdk==1.5.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.16.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.13.4
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
zipp==3.6.0
